common/sync_filesystem: make syncfs work with valgrind
Under valgrind we get ENOSYS for syncfs().  If we get that error try each
possible variant in turn, eventually resorting to vanilla sync(2).  We
may an error value in that case, but we are concerned about catching IO
errors here, not ENOSYS, and if syncfs is not supported we have already
lost that ability anyway.

Signed-off-by: Sage Weil <sage@redhat.com>